Add state.rs file with working with state logic
