+++ crates/oct-cloud/src/aws.rs
@@ -1,3 +1,4 @@
+use crate::state::{Ec2InstanceState, InstanceProfileState, InstanceRoleState};
 use aws_config;
 pub use aws_sdk_ec2;
 use aws_sdk_ec2::operation::run_instances::RunInstancesOutput;
@@ -21,34 +22,6 @@ use mockall::automock;
 pub trait Resource {
     async fn create(&mut self) -> Result<(), Box<dyn std::error::Error>>;
     async fn destroy(&mut self) -> Result<(), Box<dyn std::error::Error>>;
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Ec2InstanceState {
-    pub id: String,
-    pub arn: String,
-    pub public_ip: String,
-    pub public_dns: String,
-    pub region: String,
-    pub ami: String,
-    pub instance_type: String,
-    pub name: String,
-    pub instance_profile: Option<InstanceProfileState>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct InstanceProfileState {
-    pub name: String,
-    pub region: String,
-    pub instance_roles: Vec<InstanceRoleState>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct InstanceRoleState {
-    pub name: String,
-    pub region: String,
-    pub assume_role_policy: String,
-    pub policy_arns: Vec<String>,
 }
 
 #[derive(Debug)]
