+++ crates/oct-cloud/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod aws;
+pub mod state;
